+++ assets/js/theme.js
@@ -3,16 +3,16 @@
  */
 ( function() {
 
-    jQuery('.onepress-menu li.menu-item-has-children').each( function() {
-        jQuery(this).prepend('<div class="nav-toggle-subarrow"></div>');
-    });
-
     jQuery('#nav-toggle').on('click', function(event){
         event.preventDefault();
         jQuery('#nav-toggle').toggleClass('nav-is-visible');
         jQuery('.main-navigation .onepress-menu').toggleClass("onepress-menu-mobile");
         jQuery('.header-widget').toggleClass("header-widget-mobile");
     });
+
+    jQuery('.onepress-menu li.menu-item-has-children, .onepress-menu li.page_item_has_children').each( function() {
+		jQuery(this).prepend('<div class="nav-toggle-subarrow"><i class="fa fa-angle-down"></i></div>');
+	});
 
     jQuery('.nav-toggle-subarrow, .nav-toggle-subarrow .nav-toggle-subarrow').click(
         function () {
@@ -167,6 +167,16 @@ function res() {
 })();
 
 /**
+ * Center vertical align for navigation.
+ */
+( function() {
+    if ( onepress_js_settings.onepress_vertical_align_menu == '1' ) {
+        var header_height = jQuery('.site-header').height();
+        jQuery('.site-header .onepress-menu').css( 'line-height', header_height + "px" );
+    }
+})();
+
+/**
  * Sticky header when scroll.
  */
 ( function() {
@@ -230,6 +240,9 @@ function res() {
 
     // Smooth scroll animation
     function smoothScroll(urlhash) {
+        if ( urlhash.length <= 0 ) {
+            return false;
+        }
         jQuery("html, body").animate({
             scrollTop: (jQuery(urlhash).offset().top - h) + "px"
         }, {