+++ inc/dashboard.php
@@ -99,14 +99,15 @@ function onepress_theme_info_page() {
                             <h3><?php esc_html_e( 'Theme Documentation', 'onepress' ); ?></h3>
                             <p class="about"><?php printf(esc_html__('Need any help to setup and configure %s? Please have a look at our documentations instructions.', 'onepress'), $theme_data->Name); ?></p>
                             <p>
-                                <a href="<?php echo esc_url( 'http://docs.famethemes.com/category/42-onepress' ); ?>" target="_blank" class="button button-secondary"><?php esc_html_e('Online Documentation', 'onepress'); ?></a>
+                                <a href="<?php echo esc_url( 'http://docs.famethemes.com/category/42-onepress' ); ?>" target="_blank" class="button button-secondary"><?php esc_html_e('OnePress Documentation', 'onepress'); ?></a>
                             </p>
+                            <?php do_action( 'onepress_dashboard_theme_links' ); ?>
                         </div>
                         <div class="theme_link">
                             <h3><?php esc_html_e( 'Having Trouble, Need Support?', 'onepress' ); ?></h3>
-                            <p class="about"><?php printf(esc_html__('Support for %s WordPress theme is conducted through the WordPress free theme support forum.', 'onepress'), $theme_data->Name); ?></p>
+                            <p class="about"><?php printf(esc_html__('Support for %s WordPress theme is conducted through FameThemes support ticket system.', 'onepress'), $theme_data->Name); ?></p>
                             <p>
-                                <a href="<?php echo esc_url('https://wordpress.org/support/theme/onepress' ); ?>" target="_blank" class="button button-secondary"><?php echo sprintf( esc_html('Go To %s Support Forum', 'onepress'), $theme_data->Name); ?></a>
+                                <a href="<?php echo esc_url('https://www.famethemes.com/dashboard/tickets/' ); ?>" target="_blank" class="button button-secondary"><?php echo sprintf( esc_html('Create a support ticket', 'onepress'), $theme_data->Name); ?></a>
                             </p>
                         </div>
                     </div>